requirements.txt
# AID-Service Dependencies
# FastAPI framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0

# Logging
loguru>=0.7.0

# HTTP client (for algorithm layer)
httpx>=0.24.0

# Async file I/O
aiofiles>=23.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Type hints
python-multipart>=0.0.6
